BJS2-26621 refactored
