+++ src/main/java/school/faang/user_service/validator/recommendation/RecommendationServiceValidator.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.validator;
+package school.faang.user_service.validator.recommendation;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
